pyproject.toml
[project]
name = "posterpilot"
version = "1.0.0"
description = "Automatic Plex poster management"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.6",
    "plexapi>=4.15",
    "httpx>=0.27",
    "Pillow>=10.0",
    "pydantic>=2.0",
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = [
    "pyinstaller>=6.0",
    "pytest>=8.0",
    "ruff>=0.3",
]

[project.scripts]
posterpilot = "run:main"

[tool.ruff]
line-length = 100
target-version = "py311"
